double to float lost precision

+++ Core/include/Acts/Seeding/SeedFinder.ipp
@@ -603,7 +603,7 @@ inline void SeedFinder<external_spacepoint_t, platform_t>::filterCandidates(
         cotThetaT = zT * std::sqrt(iDeltaRT2);
 
         rMxy = std::sqrt(rMTransf[0] * rMTransf[0] + rMTransf[1] * rMTransf[1]);
-        float irMxy = 1 / rMxy;
+        double irMxy = 1 / rMxy;
         float Ax = rMTransf[0] * irMxy;
         float Ay = rMTransf[1] * irMxy;
